requirements.txt
qiskit==1.0.0
qiskit-ibm-provider==0.7.0
pennylane==0.36.0
cupy-cuda12x==12.3.0      # adjust based on your CUDA version
numpy==1.26.4